requirements.txt
dash
pandas
db-dtypes
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-secret-manager
gunicorn
Flask
Flask-OIDC
black
debugpy
pylint
pyarrow
